+++ tests/testthat/test-parquet.R
@@ -74,3 +74,13 @@ test_that("write_paquet works", {
     ignore_attr = TRUE
   )
 })
+
+test_that("throw error if invalid compression is passed", {
+  tmpf = tempfile()
+  on.exit(unlink(tmpf))
+  df_exp = pl$DataFrame(mtcars)
+  expect_error(
+    df_exp$write_parquet(tmpf, compression = "invalid"),
+    "Failed to set parquet compression method"
+  )
+})