+++ NEWS.md
@@ -91,7 +91,7 @@
         replacement of `pl$reset_options()`).
 -   New function `polars_envvars()` to print the list of environment variables
     related to polars (#735).
--   New method `$write_parquet()` for DataFrame (#745).
+-   New method `$write_parquet()` for DataFrame (#758).
 
 ## polars 0.12.2